Animations and Cyan

+++ rangers-api/rangers-api/Application/Player/EffectParticleLocation.h
@@ -9,7 +9,8 @@ namespace app::player {
         uint8_t flags;
 
         CREATE_FUNC(EffectParticleLocation, hh::game::GameObject* playerObject, PlayerVisual* playerVisual);
-        virtual unsigned long long GetNameHash() const override;
+
+        virtual unsigned int GetNameHash() const override;
         virtual void Update(int64_t a2, float deltaTime) override;
         virtual void PreCreateParticle(hh::animeff::GOCParticleLocator* gocParticleLocator, PreCreateParticleInfo& preCreateParticleInfo) override;
     };
